stocksage-api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
yfinance==0.2.51
pandas==2.2.3
requests==2.32.3
cachetools==5.5.0
python-dotenv==1.0.1
pytz==2024.1
apscheduler==3.10.4
